feat: TimerView에 카운트다운 기능 구현
